FW-1158: Reduce code duplication in tests
