+++ FirstVoicesData/src/test/java/ca/firstvoices/operations/InitialDatabaseSetupTest.java
@@ -1,83 +1,23 @@
 package ca.firstvoices.operations;
 
-import ca.firstvoices.testUtil.AbstractTest;
+import ca.firstvoices.testUtil.AbstractFirstVoicesDataTest;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.nuxeo.ecm.automation.OperationException;
-import org.nuxeo.ecm.automation.test.AutomationFeature;
-import org.nuxeo.ecm.core.api.CoreSession;
 import org.nuxeo.ecm.core.api.DocumentModel;
 import org.nuxeo.ecm.core.api.PathRef;
-import org.nuxeo.ecm.automation.AutomationService;
 import org.nuxeo.ecm.automation.OperationContext;
-import org.nuxeo.ecm.core.api.DocumentModelList;
-import org.nuxeo.ecm.core.test.CoreFeature;
-import org.nuxeo.ecm.core.test.DefaultRepositoryInit;
-import org.nuxeo.ecm.core.test.annotations.Granularity;
-import org.nuxeo.ecm.core.test.annotations.RepositoryConfig;
-import org.nuxeo.ecm.platform.test.PlatformFeature;
-import org.nuxeo.ecm.platform.usermanager.UserManager;
-import org.nuxeo.elasticsearch.test.RepositoryElasticSearchFeature;
-import org.nuxeo.runtime.test.runner.*;
-
-import javax.inject.Inject;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 
 import static org.junit.Assert.*;
-@RunWith(FeaturesRunner.class)
-@Features({AutomationFeature.class, PlatformFeature.class, RuntimeFeature.class, CoreFeature.class, RepositoryElasticSearchFeature.class})
-@RepositoryConfig(init = DefaultRepositoryInit.class, cleanup = Granularity.METHOD)
 
-@Deploy("org.nuxeo.binary.metadata")
-@Deploy("org.nuxeo.ecm.platform.url.core")
-@Deploy("org.nuxeo.ecm.platform.types.api")
-@Deploy("org.nuxeo.ecm.platform.types.core")
-@Deploy("org.nuxeo.ecm.platform.filemanager.api")
-@Deploy("org.nuxeo.ecm.platform.filemanager.core")
-@Deploy("org.nuxeo.ecm.platform.rendition.core")
-@Deploy("org.nuxeo.ecm.platform.tag")
-@Deploy("org.nuxeo.ecm.platform.commandline.executor")
-@Deploy("org.nuxeo.ecm.platform.convert")
-@Deploy("org.nuxeo.ecm.platform.preview")
-
-// Audio doctype
-@Deploy("org.nuxeo.ecm.platform.audio.core")
-
-// Video doctype
-@Deploy("org.nuxeo.ecm.platform.video.convert")
-@Deploy("org.nuxeo.ecm.platform.video.core")
-
-// Picture doctype
-@Deploy("org.nuxeo.ecm.platform.picture.core")
-@Deploy("org.nuxeo.ecm.platform.picture.api")
-@Deploy("org.nuxeo.ecm.platform.picture.convert")
-
-// ElasticSearch / Search
-@Deploy("org.nuxeo.elasticsearch.core:elasticsearch-test-contrib.xml")
-@Deploy("org.nuxeo.ecm.platform.search.core")
-@Deploy("org.nuxeo.ecm.platform.webapp.types")
-
-@PartialDeploy(bundle = "FirstVoicesData", extensions = {TargetExtensions.ContentModel.class})
-@Deploy("FirstVoicesData:OSGI-INF/ca.firstvoices.operations.xml")
-public class InitialDatabaseSetupTest extends AbstractTest {
-
-    DocumentModel sectionsRoot;
+public class InitialDatabaseSetupTest extends AbstractFirstVoicesDataTest {
 
     // Environment variables for the admin account that will be created.
     private static final String username = System.getenv("CYPRESS_FV_USERNAME");
     private static final String password = System.getenv("CYPRESS_FV_PASSWORD");
-
-    @Inject
-    private CoreSession session;
-
-    @Inject
-    private UserManager userManager;
-
-    @Inject
-    protected AutomationService automationService;
 
     @Before
     public void setUp() throws Exception {
@@ -89,9 +29,6 @@ public class InitialDatabaseSetupTest extends AbstractTest {
                 createDocument(session, session.createDocumentModel("/FV", "Workspaces", "WorkspaceRoot")));
         assertNotNull("Should have a valid SharedData directory",
                 createDocument(session, session.createDocumentModel("/FV/Workspaces", "SharedData", "Workspace")));
-
-        DocumentModelList docs = session.query("SELECT * FROM Document WHERE ecm:path = '/FV/sections'");
-        sectionsRoot = docs.get(0);
 
         assertNotNull("Should have a valid Workspaces/Data directory",
                 createDocument(session, session.createDocumentModel("/FV/Workspaces", "Data", "Workspace")));
