+++ FirstVoicesData/src/test/java/ca/firstvoices/services/CleanupCharactersServiceImplTest.java
@@ -1,71 +1,18 @@
 package ca.firstvoices.services;
 
-
 import ca.firstvoices.exceptions.FVCharacterInvalidException;
-import ca.firstvoices.testUtil.AbstractTest;
+import ca.firstvoices.testUtil.AbstractFirstVoicesDataTest;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.nuxeo.ecm.automation.test.AutomationFeature;
-import org.nuxeo.ecm.core.api.CoreSession;
 import org.nuxeo.ecm.core.api.DocumentModel;
 import org.nuxeo.ecm.core.api.DocumentModelList;
-import org.nuxeo.ecm.core.test.CoreFeature;
-import org.nuxeo.ecm.core.test.DefaultRepositoryInit;
-import org.nuxeo.ecm.core.test.annotations.Granularity;
-import org.nuxeo.ecm.core.test.annotations.RepositoryConfig;
-import org.nuxeo.ecm.platform.test.PlatformFeature;
-import org.nuxeo.elasticsearch.test.RepositoryElasticSearchFeature;
-import org.nuxeo.runtime.test.runner.*;
 
-import javax.inject.Inject;
 import java.util.*;
 
 import static org.junit.Assert.*;
 
-@RunWith(FeaturesRunner.class)
-@Features({AutomationFeature.class, PlatformFeature.class, RuntimeFeature.class, CoreFeature.class, RepositoryElasticSearchFeature.class})
-@RepositoryConfig(init = DefaultRepositoryInit.class, cleanup = Granularity.METHOD)
-
-@Deploy("org.nuxeo.binary.metadata")
-@Deploy("org.nuxeo.ecm.platform.url.core")
-@Deploy("org.nuxeo.ecm.platform.types.api")
-@Deploy("org.nuxeo.ecm.platform.types.core")
-@Deploy("org.nuxeo.ecm.platform.filemanager.api")
-@Deploy("org.nuxeo.ecm.platform.filemanager.core")
-@Deploy("org.nuxeo.ecm.platform.rendition.core")
-@Deploy("org.nuxeo.ecm.platform.tag")
-@Deploy("org.nuxeo.ecm.platform.commandline.executor")
-@Deploy("org.nuxeo.ecm.platform.convert")
-@Deploy("org.nuxeo.ecm.platform.preview")
-
-// Audio doctype
-@Deploy("org.nuxeo.ecm.platform.audio.core")
-
-// Video doctype
-@Deploy("org.nuxeo.ecm.platform.video.convert")
-@Deploy("org.nuxeo.ecm.platform.video.core")
-
-// Picture doctype
-@Deploy("org.nuxeo.ecm.platform.picture.core")
-@Deploy("org.nuxeo.ecm.platform.picture.api")
-@Deploy("org.nuxeo.ecm.platform.picture.convert")
-
-// ElasticSearch / Search
-@Deploy("org.nuxeo.elasticsearch.core:elasticsearch-test-contrib.xml")
-@Deploy("org.nuxeo.ecm.platform.search.core")
-@Deploy("org.nuxeo.ecm.platform.webapp.types")
-
-@PartialDeploy(bundle = "FirstVoicesData", extensions = {TargetExtensions.ContentModel.class})
-@Deploy("FirstVoicesData:OSGI-INF/services/ca.firstvoices.services.cleanupcharacterservice.xml")
-public class CleanupCharactersServiceImplTest extends AbstractTest {
-
-    @Inject
-    private CoreSession session;
-
-    @Inject
-    private CleanupCharactersService service;
+public class CleanupCharactersServiceImplTest extends AbstractFirstVoicesDataTest {
 
     private DocumentModel dialect;
 
@@ -94,7 +41,7 @@ public class CleanupCharactersServiceImplTest extends AbstractTest {
         List<DocumentModel> documentModels = createWords(words);
         for (int i = 0; i < documentModels.size(); i++) {
             DocumentModel documentModel = documentModels.get(i);
-            DocumentModel updatedDocument = service.cleanConfusables(documentModel);
+            DocumentModel updatedDocument = cleanupCharactersService.cleanConfusables(documentModel);
             String title = (String) updatedDocument.getPropertyValue("dc:title");
             assertEquals(correctWords[i], title);
         }
@@ -103,7 +50,7 @@ public class CleanupCharactersServiceImplTest extends AbstractTest {
     @Test
     public void mapAndValidateConfusableCharactersTest() {
         DocumentModelList characters = session.getChildren(getAlphabetDoc().getRef());
-        Map<String, String> charMap = service.mapAndValidateConfusableCharacters(characters);
+        Map<String, String> charMap = cleanupCharactersService.mapAndValidateConfusableCharacters(characters);
         assertEquals(9, charMap.size());
         for (Map.Entry<String, String> pair : charMap.entrySet()) {
             if ("∀".equals(pair.getKey())) {
@@ -137,7 +84,7 @@ public class CleanupCharactersServiceImplTest extends AbstractTest {
         alphabetAndConfusableCharacters.put("f", new String[] {"e"});
         createAlphabetWithConfusableCharacters(alphabetAndConfusableCharacters);
         DocumentModelList characters = session.getChildren(getAlphabetDoc().getRef());
-        service.mapAndValidateConfusableCharacters(characters);
+        cleanupCharactersService.mapAndValidateConfusableCharacters(characters);
     }
 
     private void createAlphabetWithConfusableCharacters(Map<String, String[]> alphabet) {
